increment in items(how will they sell the item, by piece? 1/2 ? 1/4 etc)
items instock stocksold converted to double

created order and history

created an endpoint for address addition (it is a waste of resources to make the customer update the whole customer data just to add a single address)

fixed error in saving item (unsaved transient instance)

renamed order to transaction
created checkout in TransactionService

created add transactions(meaning the transaction has been confirmed)

it updates the instock in store (i think the way of doing it is inefficient, has room for improvement)

can update the history

+++ hustle-emarket-rest/src/main/java/org/emarket/hustle/hustleemarketrest/service/ItemServiceImpl.java
@@ -85,7 +85,7 @@ public class ItemServiceImpl implements ItemService
 	{
 		try
 		{
-			if(item.isDelisted())
+			if(item.isDelisted() && item.getId() != 0)
 			{
 				basketRepository.deleteByItem(item);
 			}
